Extract SlotPicker into a shared component
